pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Demetrius Panovitch <demetrius.mp789@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "20.0a4"
Pillow = "^9.2.0"
torchvision = "^0.13.1"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
Flake8-pyproject = "^1.1.0"
flake8 = "^5.0.4"
black = "^22.10.0"
isort = "^5.10.1"

[tool.isort]
profile = "black"

[tool.flake8]
max-line-length = 88
extend-ignore = "E203"
per-file-ignores = [
    # ignores unused imports on __init__.py files
    "__init__.py: F401"
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
